Add output to DG tss solvers, add arguments to run_dg_example and read_plot_1D
