+++ sfepy/discrete/dg/examples/example_dg_common.py
@@ -3,16 +3,18 @@ import numpy as nm
 from sfepy.discrete.fem import Mesh
 from sfepy.discrete.fem.meshio import UserMeshIO
 
+from sfepy.base.base import (get_default, output, assert_,
+                             Struct, basestr, IndexedStruct)
+
 from sfepy.terms import register_term
 from sfepy.solvers import  register_solver
-from sfepy.discrete.dg.examples.run_dg_examples import get_cfl_setup
-
 
 # import various ICs
 from sfepy.discrete.dg.my_utils.inits_consts import ghump, gsmooth, left_par_q, left_cos, superic, three_step_u, \
     sawtooth_q, const_q, quadr_cub
 
 from sfepy.discrete.dg.dg_terms import AdvectDGFluxTerm
+
 
 
 # import TSSs
@@ -24,6 +26,40 @@ register_solver(TVDRK3StepSolver)
 register_solver(RK4StepSolver)
 register_solver(EulerStepSolver)
 
+def get_cfl_setup(CFL):
+
+    def setup_cfl_condition(problem):
+        """
+        Sets up CFL condition for problem ts_conf in problem
+        :param problem: discrete.problem.Problem
+        :return:
+        """
+        ts_conf = problem.ts_conf
+        mesh = problem.domain.mesh
+        dim = mesh.dim
+        approx_order = problem.fields['density'].approx_order
+        mats = problem.create_materials('a')
+        velo = problem.conf_materials['material_a__0'].values["val"]
+        max_velo = nm.max(nm.linalg.norm(velo))
+        dx = nm.min(problem.domain.mesh.cmesh.get_volumes(dim))
+        order_corr = 1. / (2 * approx_order + 1)
+        dt = dx / max_velo * CFL * order_corr
+        # time_steps_N = int((tf - t0) / dt) * 2
+        tn = int(nm.ceil((ts_conf.t1 - ts_conf.t0) / dt))
+        dtdx = dt / dx
+
+        ts_conf += Struct(dt=dt, n_step=tn)
+        output("Preprocessing hook setup_cfl_condition:")
+        output("Approximation order of field {} order is {}".format(problem.fields['density'].family_name, approx_order))
+        output("Space divided into {0} cells, {1} steps, step size is {2}".format(mesh.n_el, len(mesh.coors), dx))
+        output("Time divided into {0} nodes, {1} steps, step size is {2}".format(tn - 1, tn, dt))
+        output("CFL coefficient was {0} and order correction 1/{1} = {2}".format(CFL,  (2 * approx_order + 1), order_corr))
+        output("Courant number c = max(norm(a)) * dt/dx = {0}".format(max_velo * dtdx))
+        output("------------------------------------------")
+        output("Time stepping solver is {}".format(ts_conf.kind))
+
+
+    return setup_cfl_condition
 
 def get_common(approx_order, CFL, t0, t1, limiter, get_ic):
